Removed comments and unused imports.
